requirements.txt
Flask==3.0.0
Flask_SQLAlchemy==3.1.1
python-dotenv
psycopg2-binary
gunicorn
Werkzeug==3.0.1
Jinja2==3.1.2
